Fix people array not resetting
When the creates a new car group card and then presses plus, it
will reset the info on the car group card. However, if one changes
the first person in the people array, then whenever the user
creates a new card, it will essentially edit the value in the
defaultcargroup object, making all new default entries with the newly
entered person. So if I entered 小刘 in the first entry in the people
array, it will keep saying 小刘 whenever I press the plus button.
To solve this, I just essentially forced it to be 小冰 when
resetting the carGroup state.

setCarGroupInfo({...defaultCarGroupInfo, People: [{name: "小冰"}]});